Complete next hardening wave: OpenAPI contracts, coverage, RBAC smoke, and sync splits
