+++ .github/workflows/ci.yml
@@ -112,7 +112,7 @@ jobs:
       - name: Enforce minimum coverage
         shell: bash
         env:
-          MIN_COVERAGE: "40.6"
+          MIN_COVERAGE: "41.0"
         run: |
           set -euo pipefail
           mapfile -t all_packages < <(go list ./...)
@@ -137,12 +137,12 @@ jobs:
           from pathlib import Path
 
           required = {
-              "github.com/writer/cerebro/internal/api": 36.0,
+              "github.com/writer/cerebro/internal/api": 40.0,
               "github.com/writer/cerebro/internal/app": 47.0,
               "github.com/writer/cerebro/internal/cache": 82.0,
               "github.com/writer/cerebro/internal/graph": 62.0,
               "github.com/writer/cerebro/internal/policy": 72.0,
-              "github.com/writer/cerebro/internal/scanner": 36.0,
+              "github.com/writer/cerebro/internal/scanner": 44.0,
               "github.com/writer/cerebro/internal/scheduler": 71.5,
           }
 
@@ -305,6 +305,97 @@ jobs:
           if [ "$xkey_status" -ne "200" ]; then
             echo "expected 200 for X-API-Key request, got $xkey_status"
             cat /tmp/cerebro-api-auth-smoke.log || true
+            exit 1
+          fi
+
+  api-auth-rbac-smoke:
+    name: api-auth-rbac-smoke
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Go
+        uses: actions/setup-go@v5
+        with:
+          go-version: '1.25.7'
+
+      - name: Build
+        run: go build -o cerebro ./cmd/cerebro
+
+      - name: RBAC role matrix smoke
+        shell: bash
+        env:
+          LOG_LEVEL: error
+          API_PORT: "18081"
+          API_AUTH_ENABLED: "true"
+          API_KEYS: "viewer-key:viewer-1,analyst-key:analyst-1,admin-key:admin-1"
+          RBAC_STATE_FILE: "/tmp/cerebro-rbac-role-smoke.json"
+        run: |
+          set -euo pipefail
+          cat > "$RBAC_STATE_FILE" <<'JSON'
+          {
+            "users": {
+              "viewer-1": {
+                "id": "viewer-1",
+                "email": "viewer@example.com",
+                "name": "Viewer",
+                "role_ids": ["viewer"]
+              },
+              "analyst-1": {
+                "id": "analyst-1",
+                "email": "analyst@example.com",
+                "name": "Analyst",
+                "role_ids": ["analyst"]
+              },
+              "admin-1": {
+                "id": "admin-1",
+                "email": "admin@example.com",
+                "name": "Admin",
+                "role_ids": ["admin"]
+              }
+            }
+          }
+          JSON
+
+          ./cerebro serve > /tmp/cerebro-api-auth-rbac-smoke.log 2>&1 &
+          server_pid=$!
+          cleanup() {
+            kill "$server_pid" >/dev/null 2>&1 || true
+            wait "$server_pid" >/dev/null 2>&1 || true
+          }
+          trap cleanup EXIT
+
+          for i in $(seq 1 45); do
+            if curl -fsS "http://127.0.0.1:${API_PORT}/health" >/dev/null 2>&1; then
+              break
+            fi
+            sleep 1
+          done
+          curl -fsS "http://127.0.0.1:${API_PORT}/health" >/dev/null
+
+          endpoint="http://127.0.0.1:${API_PORT}/api/v1/agents/sessions"
+          payload='{"agent_id":"missing-agent"}'
+
+          viewer_status=$(curl -sS -H "Authorization: Bearer viewer-key" -H "Content-Type: application/json" \
+            -d "$payload" -o /tmp/api-auth-rbac-viewer.json -w "%{http_code}" "$endpoint")
+          analyst_status=$(curl -sS -H "Authorization: Bearer analyst-key" -H "Content-Type: application/json" \
+            -d "$payload" -o /tmp/api-auth-rbac-analyst.json -w "%{http_code}" "$endpoint")
+          admin_status=$(curl -sS -H "Authorization: Bearer admin-key" -H "Content-Type: application/json" \
+            -d "$payload" -o /tmp/api-auth-rbac-admin.json -w "%{http_code}" "$endpoint")
+
+          if [ "$viewer_status" != "403" ]; then
+            echo "expected viewer to receive 403, got $viewer_status"
+            cat /tmp/cerebro-api-auth-rbac-smoke.log || true
+            exit 1
+          fi
+          if [ "$analyst_status" = "403" ]; then
+            echo "expected analyst to pass RBAC (non-403), got $analyst_status"
+            cat /tmp/cerebro-api-auth-rbac-smoke.log || true
+            exit 1
+          fi
+          if [ "$admin_status" = "403" ]; then
+            echo "expected admin to pass RBAC (non-403), got $admin_status"
+            cat /tmp/cerebro-api-auth-rbac-smoke.log || true
             exit 1
           fi
