+++ internal/sync/relationships_aws_core.go
@@ -1,0 +1,502 @@
+package sync
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"strings"
+)
+
+func (r *RelationshipExtractor) extractEC2Relationships(ctx context.Context) (int, error) {
+	query := `SELECT ARN, ACCOUNT_ID, REGION, VPC_ID, SUBNET_ID, IAM_INSTANCE_PROFILE, SECURITY_GROUPS 
+	          FROM AWS_EC2_INSTANCES WHERE ARN IS NOT NULL`
+
+	result, err := r.sf.Query(ctx, query)
+	if err != nil {
+		if isMissingRelationshipSourceError(err) {
+			return 0, nil
+		}
+		return 0, err
+	}
+
+	var rels []Relationship
+	for _, row := range result.Rows {
+		instanceARN := toString(queryRow(row, "arn"))
+		if instanceARN == "" {
+			continue
+		}
+		accountID := toString(queryRow(row, "account_id"))
+		region := toString(queryRow(row, "region"))
+
+		// VPC relationship
+		if vpcID := toString(queryRow(row, "vpc_id")); vpcID != "" {
+			vpcARN := awsARNForResource("vpc", region, accountID, vpcID)
+			rels = append(rels, Relationship{
+				SourceID:   instanceARN,
+				SourceType: "aws:ec2:instance",
+				TargetID:   vpcARN,
+				TargetType: "aws:ec2:vpc",
+				RelType:    RelInVPC,
+			})
+		}
+
+		// Subnet relationship
+		if subnetID := toString(queryRow(row, "subnet_id")); subnetID != "" {
+			subnetARN := awsARNForResource("subnet", region, accountID, subnetID)
+			rels = append(rels, Relationship{
+				SourceID:   instanceARN,
+				SourceType: "aws:ec2:instance",
+				TargetID:   subnetARN,
+				TargetType: "aws:ec2:subnet",
+				RelType:    RelInSubnet,
+			})
+		}
+
+		// IAM instance profile relationship
+		if profile := queryRow(row, "iam_instance_profile"); profile != nil {
+			switch val := profile.(type) {
+			case map[string]interface{}:
+				if roleARN := toString(val["arn"]); roleARN != "" {
+					rels = append(rels, Relationship{
+						SourceID:   instanceARN,
+						SourceType: "aws:ec2:instance",
+						TargetID:   roleARN,
+						TargetType: "aws:iam:instance_profile",
+						RelType:    RelHasRole,
+					})
+				} else if roleARN := toString(val["Arn"]); roleARN != "" {
+					rels = append(rels, Relationship{
+						SourceID:   instanceARN,
+						SourceType: "aws:ec2:instance",
+						TargetID:   roleARN,
+						TargetType: "aws:iam:instance_profile",
+						RelType:    RelHasRole,
+					})
+				}
+			case string:
+				if val != "" {
+					rels = append(rels, Relationship{
+						SourceID:   instanceARN,
+						SourceType: "aws:ec2:instance",
+						TargetID:   val,
+						TargetType: "aws:iam:instance_profile",
+						RelType:    RelHasRole,
+					})
+				}
+			}
+		}
+
+		// Security group relationships
+		if sgList := asSlice(queryRow(row, "security_groups")); len(sgList) > 0 {
+			for _, sg := range sgList {
+				if sgMap := asMap(sg); sgMap != nil {
+					if sgID := getStringAny(sgMap, "GroupId", "groupId", "group_id"); sgID != "" {
+						sgARN := awsARNForResource("security-group", region, accountID, sgID)
+						rels = append(rels, Relationship{
+							SourceID:   instanceARN,
+							SourceType: "aws:ec2:instance",
+							TargetID:   sgARN,
+							TargetType: "aws:ec2:security_group",
+							RelType:    RelMemberOf,
+						})
+					}
+				}
+			}
+		}
+	}
+
+	return r.persistRelationships(ctx, rels)
+}
+
+func (r *RelationshipExtractor) extractIAMRoleRelationships(ctx context.Context) (int, error) {
+	query := `SELECT ARN, ROLE_NAME, ASSUME_ROLE_POLICY_DOCUMENT 
+	          FROM AWS_IAM_ROLES WHERE ARN IS NOT NULL`
+
+	result, err := r.sf.Query(ctx, query)
+	if err != nil {
+		if isMissingRelationshipSourceError(err) {
+			return 0, nil
+		}
+		return 0, err
+	}
+
+	var rels []Relationship
+	for _, row := range result.Rows {
+		roleARN := toString(queryRow(row, "arn"))
+		if roleARN == "" {
+			continue
+		}
+
+		// Parse trust policy to extract who can assume the role
+		if trustPolicy := queryRow(row, "assume_role_policy_document"); trustPolicy != nil {
+			policyDoc, err := parsePolicyDocument(trustPolicy)
+			if err != nil {
+				r.logger.Warn("failed to parse trust policy", "role", roleARN, "error", err)
+				continue
+			}
+			if policyDoc == nil {
+				continue
+			}
+			if statements, ok := policyDoc["Statement"].([]interface{}); ok {
+				for _, stmt := range statements {
+					if stmtMap, ok := stmt.(map[string]interface{}); ok {
+						if effect := toString(stmtMap["Effect"]); effect == "Allow" {
+							if principal := stmtMap["Principal"]; principal != nil {
+								principals := extractPrincipals(principal)
+								for _, p := range principals {
+									rels = append(rels, Relationship{
+										SourceID:   roleARN,
+										SourceType: "aws:iam:role",
+										TargetID:   p,
+										TargetType: inferPrincipalType(p),
+										RelType:    RelAssumableBy,
+									})
+								}
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+
+	return r.persistRelationships(ctx, rels)
+}
+
+func (r *RelationshipExtractor) extractLambdaRelationships(ctx context.Context) (int, error) {
+	query := `SELECT ARN, FUNCTION_NAME, ROLE, VPC_CONFIG 
+	          FROM AWS_LAMBDA_FUNCTIONS WHERE ARN IS NOT NULL`
+
+	result, err := r.sf.Query(ctx, query)
+	if err != nil {
+		if isMissingRelationshipSourceError(err) {
+			return 0, nil
+		}
+		return 0, err
+	}
+
+	var rels []Relationship
+	for _, row := range result.Rows {
+		functionARN := toString(queryRow(row, "arn"))
+		if functionARN == "" {
+			continue
+		}
+		region, accountID := awsRegionAccountFromARN(functionARN)
+
+		// Execution role relationship
+		if roleARN := toString(queryRow(row, "role")); roleARN != "" {
+			rels = append(rels, Relationship{
+				SourceID:   functionARN,
+				SourceType: "aws:lambda:function",
+				TargetID:   roleARN,
+				TargetType: "aws:iam:role",
+				RelType:    RelHasRole,
+			})
+		}
+
+		// VPC relationships
+		if vpcConfig := asMap(queryRow(row, "vpc_config")); vpcConfig != nil {
+			if vpcID := getStringAny(vpcConfig, "VpcId", "vpcId", "vpc_id"); vpcID != "" {
+				vpcARN := awsARNForResource("vpc", region, accountID, vpcID)
+				rels = append(rels, Relationship{
+					SourceID:   functionARN,
+					SourceType: "aws:lambda:function",
+					TargetID:   vpcARN,
+					TargetType: "aws:ec2:vpc",
+					RelType:    RelInVPC,
+				})
+			}
+
+			// Security groups
+			if sgs := asSlice(vpcConfig["SecurityGroupIds"]); len(sgs) > 0 {
+				for _, sg := range sgs {
+					if sgID := toString(sg); sgID != "" {
+						sgARN := awsARNForResource("security-group", region, accountID, sgID)
+						rels = append(rels, Relationship{
+							SourceID:   functionARN,
+							SourceType: "aws:lambda:function",
+							TargetID:   sgARN,
+							TargetType: "aws:ec2:security_group",
+							RelType:    RelMemberOf,
+						})
+					}
+				}
+			}
+		}
+	}
+
+	return r.persistRelationships(ctx, rels)
+}
+
+func (r *RelationshipExtractor) extractSecurityGroupRelationships(ctx context.Context) (int, error) {
+	query := `SELECT ARN, ACCOUNT_ID, REGION, GROUP_ID, VPC_ID, IP_PERMISSIONS, IP_PERMISSIONS_EGRESS 
+	          FROM AWS_EC2_SECURITY_GROUPS WHERE ARN IS NOT NULL`
+
+	result, err := r.sf.Query(ctx, query)
+	if err != nil {
+		if isMissingRelationshipSourceError(err) {
+			return 0, nil
+		}
+		return 0, err
+	}
+
+	var rels []Relationship
+	for _, row := range result.Rows {
+		sgARN := toString(queryRow(row, "arn"))
+		if sgARN == "" {
+			continue
+		}
+		accountID := toString(queryRow(row, "account_id"))
+		region := toString(queryRow(row, "region"))
+
+		// VPC relationship
+		if vpcID := toString(queryRow(row, "vpc_id")); vpcID != "" {
+			vpcARN := awsARNForResource("vpc", region, accountID, vpcID)
+			rels = append(rels, Relationship{
+				SourceID:   sgARN,
+				SourceType: "aws:ec2:security_group",
+				TargetID:   vpcARN,
+				TargetType: "aws:ec2:vpc",
+				RelType:    RelBelongsTo,
+			})
+		}
+
+		// Check for internet exposure (0.0.0.0/0 ingress)
+		if permList := asSlice(queryRow(row, "ip_permissions")); len(permList) > 0 {
+			for _, perm := range permList {
+				permMap := asMap(perm)
+				if permMap == nil {
+					continue
+				}
+				if ranges := asSlice(permMap["IpRanges"]); len(ranges) > 0 {
+					for _, r := range ranges {
+						rMap := asMap(r)
+						if rMap == nil {
+							continue
+						}
+						cidr := toString(rMap["CidrIp"])
+						if cidr == "0.0.0.0/0" || cidr == "::/0" {
+							props, _ := encodeProperties(map[string]interface{}{
+								"from_port": permMap["FromPort"],
+								"to_port":   permMap["ToPort"],
+								"protocol":  permMap["IpProtocol"],
+								"cidr":      cidr,
+							})
+							rels = append(rels, Relationship{
+								SourceID:   sgARN,
+								SourceType: "aws:ec2:security_group",
+								TargetID:   "internet",
+								TargetType: "network:internet",
+								RelType:    RelExposedTo,
+								Properties: props,
+							})
+						}
+					}
+				}
+			}
+		}
+	}
+
+	return r.persistRelationships(ctx, rels)
+}
+
+func (r *RelationshipExtractor) extractS3Relationships(ctx context.Context) (int, error) {
+	query := `SELECT ARN, NAME, ENCRYPTION, LOGGING_TARGET_BUCKET
+	          FROM AWS_S3_BUCKETS WHERE ARN IS NOT NULL`
+
+	result, err := r.sf.Query(ctx, query)
+	if err != nil {
+		if isMissingRelationshipSourceError(err) {
+			return 0, nil
+		}
+		return 0, err
+	}
+
+	var rels []Relationship
+	for _, row := range result.Rows {
+		bucketARN := toString(queryRow(row, "arn"))
+		if bucketARN == "" {
+			continue
+		}
+
+		// KMS encryption relationship - extract from ENCRYPTION column
+		if enc := queryRow(row, "encryption"); enc != nil {
+			encStr := toString(enc)
+			// Check if KMS encryption is configured
+			if strings.Contains(encStr, "aws:kms") || strings.Contains(encStr, "KMS") {
+				// Try to parse as JSON to extract KMS key ARN
+				var encMap map[string]interface{}
+				if err := json.Unmarshal([]byte(encStr), &encMap); err == nil {
+					if kmsKeyID := toString(encMap["KMSMasterKeyID"]); kmsKeyID != "" {
+						rels = append(rels, Relationship{
+							SourceID:   bucketARN,
+							SourceType: "aws:s3:bucket",
+							TargetID:   kmsKeyID,
+							TargetType: "aws:kms:key",
+							RelType:    RelEncryptedBy,
+						})
+					}
+				}
+			}
+		}
+
+		// Logging relationship
+		if targetBucket := toString(queryRow(row, "logging_target_bucket")); targetBucket != "" {
+			rels = append(rels, Relationship{
+				SourceID:   bucketARN,
+				SourceType: "aws:s3:bucket",
+				TargetID:   fmt.Sprintf("arn:aws:s3:::%s", targetBucket),
+				TargetType: "aws:s3:bucket",
+				RelType:    RelLogsTo,
+			})
+		}
+	}
+
+	return r.persistRelationships(ctx, rels)
+}
+
+func (r *RelationshipExtractor) extractECSRelationships(ctx context.Context) (int, error) {
+	query := `SELECT ARN, CLUSTER_ARN, TASK_DEFINITION, NETWORK_CONFIGURATION
+	          FROM AWS_ECS_SERVICES WHERE ARN IS NOT NULL`
+
+	result, err := r.sf.Query(ctx, query)
+	if err != nil {
+		if isMissingRelationshipSourceError(err) {
+			return 0, nil
+		}
+		return 0, err
+	}
+
+	var rels []Relationship
+	for _, row := range result.Rows {
+		serviceARN := toString(queryRow(row, "arn"))
+
+		// Cluster relationship
+		if clusterARN := toString(queryRow(row, "cluster_arn")); clusterARN != "" {
+			rels = append(rels, Relationship{
+				SourceID:   serviceARN,
+				SourceType: "aws:ecs:service",
+				TargetID:   clusterARN,
+				TargetType: "aws:ecs:cluster",
+				RelType:    RelBelongsTo,
+			})
+		}
+
+		// Task definition relationship
+		if taskDef := toString(queryRow(row, "task_definition")); taskDef != "" {
+			rels = append(rels, Relationship{
+				SourceID:   serviceARN,
+				SourceType: "aws:ecs:service",
+				TargetID:   taskDef,
+				TargetType: "aws:ecs:task_definition",
+				RelType:    RelManagedBy,
+			})
+		}
+	}
+
+	return r.persistRelationships(ctx, rels)
+}
+
+func recordRDSLookup(lookup map[string]string, region, identifier, arn string) {
+	if identifier == "" || arn == "" {
+		return
+	}
+	if region != "" {
+		lookup[rdsLookupKey(region, identifier)] = arn
+	}
+	fallbackKey := rdsLookupKey("", identifier)
+	if _, exists := lookup[fallbackKey]; !exists {
+		lookup[fallbackKey] = arn
+	}
+}
+
+func lookupRDSResourceARN(lookup map[string]string, region, identifier string) string {
+	if identifier == "" {
+		return ""
+	}
+	if region != "" {
+		if arn, ok := lookup[rdsLookupKey(region, identifier)]; ok {
+			return arn
+		}
+	}
+	return lookup[rdsLookupKey("", identifier)]
+}
+
+func rdsLookupKey(region, identifier string) string {
+	return strings.ToLower(strings.TrimSpace(region)) + "|" + strings.ToLower(strings.TrimSpace(identifier))
+}
+
+func resolveRDSEventSourceTarget(sourceType, region, accountID, sourceID string, lookupBySourceType map[string]map[string]string) (string, string) {
+	resource, targetType := rdsEventSourceDescriptor(sourceType)
+	if sourceID == "" || targetType == "" {
+		return "", ""
+	}
+
+	normalizedType := strings.ToLower(strings.TrimSpace(sourceType))
+	if lookup := lookupBySourceType[normalizedType]; len(lookup) > 0 {
+		if targetARN := lookupRDSResourceARN(lookup, region, sourceID); targetARN != "" {
+			return targetARN, targetType
+		}
+	}
+
+	return awsRDSARN(resource, region, accountID, sourceID), targetType
+}
+
+func rdsEventSourceDescriptor(sourceType string) (resource, targetType string) {
+	switch strings.ToLower(strings.TrimSpace(sourceType)) {
+	case "db-instance":
+		return "db", "aws:rds:db_instance"
+	case "db-cluster":
+		return "cluster", "aws:rds:db_cluster"
+	case "db-snapshot":
+		return "snapshot", "aws:rds:db_snapshot"
+	case "db-cluster-snapshot":
+		return "cluster-snapshot", "aws:rds:db_cluster_snapshot"
+	case "db-subnet-group":
+		return "subgrp", "aws:rds:db_subnet_group"
+	case "db-parameter-group":
+		return "pg", "aws:rds:db_parameter_group"
+	case "db-cluster-parameter-group":
+		return "cluster-pg", "aws:rds:db_cluster_parameter_group"
+	case "db-option-group":
+		return "og", "aws:rds:db_option_group"
+	case "db-proxy":
+		return "db-proxy", "aws:rds:db_proxy"
+	case "db-security-group":
+		return "secgrp", "aws:rds:db_security_group"
+	default:
+		return "", ""
+	}
+}
+
+func awsARNForResource(resource string, region, accountID, id string) string {
+	if region == "" || accountID == "" || id == "" {
+		return id
+	}
+	return fmt.Sprintf("arn:aws:ec2:%s:%s:%s/%s", region, accountID, resource, id)
+}
+
+func awsRDSARN(resource, region, accountID, identifier string) string {
+	if identifier == "" {
+		return ""
+	}
+	if region == "" || accountID == "" {
+		return identifier
+	}
+	return fmt.Sprintf("arn:aws:rds:%s:%s:%s:%s", region, accountID, resource, identifier)
+}
+
+func awsEKSClusterARN(region, accountID, clusterName string) string {
+	if region == "" || accountID == "" || clusterName == "" {
+		return ""
+	}
+	return fmt.Sprintf("arn:aws:eks:%s:%s:cluster/%s", region, accountID, clusterName)
+}
+
+func awsRegionAccountFromARN(arn string) (string, string) {
+	parts := strings.Split(arn, ":")
+	if len(parts) < 6 {
+		return "", ""
+	}
+	return parts[3], parts[4]
+}
